backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator>=2.1.0.post1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # For PostgreSQL
# pymysql>=1.1.0  # For MySQL
# sqlite3  # Built-in for SQLite

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
pydantic==2.5.2
pydantic-settings>=2.1.0
requests==2.31.0
httpx>=0.26.0
python-dateutil>=2.8.2
psutil==5.9.7  # For system monitoring
watchfiles>=0.21.0  # For auto-reload

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.26.0

# Development
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0

# Logging & Monitoring
structlog>=24.1.0
python-json-logger>=2.0.7

# Optional Dependencies
# zerotier==0.1.0  # Not available in PyPI, will use manual implementation
# aiohttp>=3.9.3  # For async HTTP requests
# redis>=5.0.1  # For caching/rate limiting
# prometheus-client>=0.19.0  # For metrics

# New dependencies
websockets==12.0
# python-iptables==1.0.0  # Not available for Windows

# ML Model dependencies
joblib>=1.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0 
